refactor: Move custom hooks to a separate folder
