+++ src/hooks/useSpecies.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query'
-import { getSpecies } from './getSpecies'
+import { getSpecies } from '../api/getSpecies'
 
 export const useGetSpecies = (pokeId: number) => {
   return useQuery({
