+++ src/hooks/useTypes.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query'
-import { getTypes } from './getTypes'
+import { getTypes } from '../api/getTypes'
 
 export const useGetTypes = (id: number | string) => {
   return useQuery({
