+++ src/pages/home/Index.tsx
@@ -1,8 +1,8 @@
 import { Col, Flex, Row, theme } from 'antd'
 import { Content } from 'antd/es/layout/layout'
 import { useState } from 'react'
-import { useGetAllPokemon, useGetPokemon } from '../../api/usePokemon'
 import Logo from '../../assets/Logo'
+import { useGetAllPokemon, useGetPokemon } from '../../hooks/usePokemon'
 import type { PokemonProps } from '../../types/pokemonTypes'
 import PokeGrid from './components/PokeGrid'
 import Searchbar from './components/Searchbar'
