+++ src/pages/pokemon-detail/Index.tsx
@@ -2,11 +2,11 @@ import { ArrowLeftOutlined } from '@ant-design/icons'
 import { Button, Col, Flex, Row, theme } from 'antd'
 import { Content } from 'antd/es/layout/layout'
 import { Link, useNavigate, useParams } from 'react-router'
-import { useGetEvolutionChain } from '../../api/useEvolutionChain'
-import { useGetPokemon } from '../../api/usePokemon'
-import { useGetSpecies } from '../../api/useSpecies'
-import { useGetTypes } from '../../api/useTypes'
 import Logo from '../../assets/Logo'
+import { useGetEvolutionChain } from '../../hooks/useEvolutionChain'
+import { useGetPokemon } from '../../hooks/usePokemon'
+import { useGetSpecies } from '../../hooks/useSpecies'
+import { useGetTypes } from '../../hooks/useTypes'
 import EvolutionChain from './components/EvolutionChain'
 import PokeEffectiveness from './components/PokeEffectiveness'
 import PokeInfos from './components/PokeInfos'
